requirements.txt
torch
torchvision
flask
flask-cors
Pillow
numpy
matplotlib
seaborn
jupyter
